+++ Xamarin.Forms.NeoControls/NeoFrame.cs
@@ -7,16 +7,8 @@ using Xamarin.Forms.NeoControls.Extensions;
 
 namespace Xamarin.Forms.NeoControls
 {
-    public class NeoFrame : NeoView
+    public class NeoFrame : NeoRoundedView
     {
-        private const int DEFAULT_CORNER_RADIUS = 3;
-
-        public static readonly BindableProperty CornerRadiusProperty = BindableProperty.Create(
-            propertyName: nameof(CornerRadius),
-            returnType: typeof(CornerRadius),
-            declaringType: typeof(NeoButton),
-            defaultValue: new CornerRadius(DEFAULT_CORNER_RADIUS),
-            propertyChanged: OnVisualPropertyChanged);
 
         public static readonly BindableProperty ShowOutProperty = BindableProperty.Create(
             propertyName: nameof(ShowOut),
@@ -51,12 +43,6 @@ namespace Xamarin.Forms.NeoControls
             returnType: typeof(double),
             declaringType: typeof(NeoView),
             defaultValue: 1.0);
-
-        public CornerRadius CornerRadius
-        {
-            get => (CornerRadius)GetValue(CornerRadiusProperty);
-            set => SetValue(CornerRadiusProperty, value);
-        }
 
         public bool IsSoft
         {
@@ -149,45 +135,6 @@ namespace Xamarin.Forms.NeoControls
             }
 
         }
-
-        protected virtual SKPath CreatePath(float retangleWidth, float retangleHeight, float drawPadding)
-        {
-            var path = new SKPath();
-            var fTopLeftRadius = Convert.ToSingle(CornerRadius.TopLeft);
-            var fTopRightRadius = Convert.ToSingle(CornerRadius.TopRight);
-            var fBottomLeftRadius = Convert.ToSingle(CornerRadius.BottomLeft);
-            var fBottomRightRadius = Convert.ToSingle(CornerRadius.BottomRight);
-
-            var startX = fTopLeftRadius + drawPadding;
-            var startY = drawPadding;
-
-            path.MoveTo(startX, startY);
-
-            path.LineTo(retangleWidth - fTopRightRadius + drawPadding, startY);
-            path.ArcTo(fTopRightRadius,
-                new SKPoint(retangleWidth + drawPadding, fTopRightRadius + drawPadding));
-
-            path.LineTo(retangleWidth + drawPadding, retangleHeight - fBottomRightRadius + drawPadding);
-            path.ArcTo(fBottomRightRadius,
-                 new SKPoint(retangleWidth - fBottomRightRadius + drawPadding, retangleHeight + drawPadding));
-
-            path.LineTo(fBottomLeftRadius + drawPadding, retangleHeight + drawPadding);
-            path.ArcTo(fBottomLeftRadius,
-                new SKPoint(drawPadding, retangleHeight - fBottomLeftRadius + drawPadding));
-
-            path.LineTo(drawPadding, fTopLeftRadius + drawPadding);
-            path.ArcTo(fTopLeftRadius, new SKPoint(startX, startY));
-
-            path.Close();
-
-            return path;
-        }
-
-
-
-
-
-
 
     }
 }
